+++ BambooTracker/gui/instrument_editor/instrument_editor_adpcm_form.hpp
@@ -31,7 +31,6 @@ public:
 signals:
 	void jamKeyOnEvent(JamKey key);
 	void jamKeyOffEvent(JamKey key);
-	void octaveChanged(bool upFlag);
 	void modified();
 	/// 0: play song
 	/// 1: play from start